Finished all the "easy" generalization in builders
