+++ explainaboard/builders/__init__.py
@@ -1,11 +1,11 @@
 from typing import Iterable, Optional
-from explainaboard.info import SysOutputInfo, Table, Performance
+from explainaboard.info import SysOutputInfo, Table, Performance, BucketPerformance
 from explainaboard.metric import Accuracy, F1score
 import copy
 from tqdm import tqdm
 from eaas import Config, Client
 from datalabs import load_dataset
-from explainaboard.utils.analysis import *  # noqa
+from explainaboard.utils.analysis import *
 from explainaboard.utils.db_api import *
 
 
@@ -97,6 +97,22 @@ class ExplainaboardBuilder:
             )  # The config you have created above
         return self._eaas_client
 
+    def _get_true_label(self, data_point):
+        """
+        Get the true label from a data point. Returns "true_label" by default, but can be overloaded.
+        :param data_point: the data point under consideration
+        :return: the true label for the output
+        """
+        return data_point["true_label"]
+
+    def _get_predicted_label(self, data_point):
+        """
+        Get the predicted label from a data point. Returns "predicted_label" by default, but can be overloaded.
+        :param data_point: the data point under consideration
+        :return: the predicted label for the output
+        """
+        return data_point["predicted_label"]
+
     def _complete_feature(self):
         """
         This function is used to calculate features used for bucekting, such as sentence_length
@@ -136,8 +152,8 @@ class ExplainaboardBuilder:
 
         for _id, feature_table in self._data.items():
 
-            predicted_labels.append(feature_table["predicted_label"])
-            true_labels.append(feature_table["true_label"])
+            predicted_labels.append(self._get_predicted_label(feature_table))
+            true_labels.append(self._get_true_label(feature_table))
 
         for metric_name in self._info.metric_names:
             # TODO(gneubig): we should try to get rid of these "eval()" calls, as they're dangerous
@@ -163,6 +179,70 @@ class ExplainaboardBuilder:
                 self._info.results.overall[metric_name] = overall_performance
             else:
                 self._info.results.overall[metric_name] = overall_performance
+
+    def get_bucket_performance(self, feature_name: str):
+        """
+        This function defines how to get bucket-level performance w.r.t a given feature (e.g., sentence length)
+        :param feature_name: the name of a feature, e.g., sentence length
+        :return: bucket_name_to_performance: a dictionary that maps bucket names to bucket performance
+        """
+
+        bucket_name_to_performance = {}
+        for bucket_interval, sample_ids in self._samples_over_bucket[
+            feature_name
+        ].items():
+
+            bucket_true_labels = []
+            bucket_predicted_labels = []
+            bucket_cases = []
+
+            for sample_id in sample_ids:
+
+                true_label = self._get_true_label(self._data[int(sample_id)])
+                predicted_label = self._get_predicted_label(self._data[int(sample_id)])
+                s_id = self._data[int(sample_id)]["id"]
+
+                # get a bucket of true/predicted labels
+                bucket_true_labels.append(true_label)
+                bucket_predicted_labels.append(predicted_label)
+                # get a bucket of cases (e.g., errors)
+                if self._info.results.is_print_case:
+                    if true_label != predicted_label:
+                        bucket_case = str(s_id)
+                        bucket_cases.append(bucket_case)
+
+            bucket_name_to_performance[bucket_interval] = []
+            for metric_name in self._info.metric_names:
+                one_metric = eval(metric_name)(
+                    true_labels=bucket_true_labels,
+                    predicted_labels=bucket_predicted_labels,
+                    is_print_confidence_interval=self._info.results.is_print_confidence_interval,
+                )
+                bucket_value_json = one_metric.evaluate()
+
+                bucket_value = bucket_value_json["value"]
+                confidence_score_low = bucket_value_json["confidence_score_low"]
+                confidence_score_up = bucket_value_json["confidence_score_up"]
+
+                # print(f"name:\t {one_metric._name} \n"
+                #       f"value:\t {bucket_value}\n"
+                #       f"confidence low\t {confidence_score_low}\n"
+                #       f"confidence up \t {confidence_score_up}\n"
+                #       f"---------------------------------")
+
+                bucket_performance = BucketPerformance(
+                    bucket_name=bucket_interval,
+                    metric_name=metric_name,
+                    value=format(bucket_value, '.4g'),
+                    confidence_score_low=format(confidence_score_low, '.4g'),
+                    confidence_score_up=format(confidence_score_up, '.4g'),
+                    n_samples=len(bucket_true_labels),
+                    bucket_samples=bucket_cases,
+                )
+
+                bucket_name_to_performance[bucket_interval].append(bucket_performance)
+
+        return sort_dict(bucket_name_to_performance)
 
     def _bucketing_samples(self, sysout_iterator):
 
@@ -209,7 +289,7 @@ class ExplainaboardBuilder:
         for feature_name, metadata in self._performances_over_bucket.items():
             dict_fine_grained[feature_name] = []
             for bucket_name, bucket_performance in metadata.items():
-                bucket_name = beautify_interval(bucket_name)  # noqa
+                bucket_name = beautify_interval(bucket_name)
 
                 # instantiation
                 dict_fine_grained[feature_name].append(bucket_performance)
@@ -218,7 +298,7 @@ class ExplainaboardBuilder:
 
     def _print_bucket_info(self):
         for feature_name in self._performances_over_bucket.keys():
-            print_dict(  # noqa
+            print_dict(
                 self._performances_over_bucket[feature_name], feature_name
             )
