+++ explainaboard/builders/conditional_generation.py
@@ -271,6 +271,7 @@ class CondGenExplainaboardBuilder(ExplainaboardBuilder):
             else:
                 self._info.results.overall[metric_name] = overall_performance
 
+    # TODO(gneubig): this should be generalized
     def get_bucket_performance(self, feature_name: str):
         """
         This function defines how to get bucket-level performance w.r.t a given feature (e.g., sentence length)
