+++ explainaboard/builders/extractive_qa.py
@@ -170,6 +170,7 @@ class QAExtractiveExplainaboardBuilder(ExplainaboardBuilder):
             else:
                 self._info.results.overall[metric_name] = overall_performance
 
+    # TODO(gneubig): this should be generalized
     def get_bucket_performance(self, feature_name: str):
         """
         This function defines how to get bucket-level performance w.r.t a given feature (e.g., sentence length)
