+++ explainaboard/builders/named_entity_recognition.py
@@ -623,6 +623,7 @@ class NERExplainaboardBuilder(ExplainaboardBuilder):
 
         return errorCase_list
 
+    # TODO(gneubig): this may be able to be generalized
     def get_bucket_performance(self, feature_name: str):
         """
         This function defines how to get bucket-level performance w.r.t a given feature (e.g., sentence length)
